+++ ChangeLog.md
@@ -1,5 +1,7 @@
-## v4.11.11 Changes
+## v4.11.13 Changes
 
-* [Retail] Fixed the mailing issue from the previous release
+* [Classic/BCC] Added Bank gathering source
+* [BCC] Implemented Guild Bank support
+* Some crafting/inventory optimisations
 
 [Known Issues](https://support.tradeskillmaster.com/en_US/known_issues)